Do not process commands if you are not the coordinator
