+++ stacks-signer/src/runloop.rs
@@ -22,7 +22,6 @@ use hashbrown::{HashMap, HashSet};
 use libsigner::{SignerEvent, SignerRunLoop};
 use slog::{slog_debug, slog_error, slog_info, slog_warn};
 use stacks_common::types::chainstate::{StacksAddress, StacksPublicKey};
-use stacks_common::util::hash::Sha256Sum;
 use stacks_common::{debug, error, info, warn};
 use wsts::curve::ecdsa;
 use wsts::curve::point::{Compressed, Point};
@@ -317,160 +316,5 @@ impl SignerRunLoop<Vec<OperationResult>, RunLoopCommand> for RunLoop {
         // Cleanup any stale signers
         self.cleanup_stale_signers();
         None
-    }
-}
-
-/// Helper function for determining the coordinator public key given the the public keys
-pub fn calculate_coordinator(
-    public_keys: &PublicKeys,
-    stacks_client: &StacksClient,
-) -> (u32, ecdsa::PublicKey) {
-    let stacks_tip_consensus_hash = match stacks_client.get_stacks_tip_consensus_hash() {
-        Ok(hash) => hash,
-        Err(e) => {
-            error!("Error in fetching consensus hash: {:?}", e);
-            return (0, public_keys.signers.get(&0).cloned().unwrap());
-        }
-    };
-    debug!(
-        "Using stacks_tip_consensus_hash {:?} for selecting coordinator",
-        &stacks_tip_consensus_hash
-    );
-
-    // Create combined hash of each signer's public key with stacks_tip_consensus_hash
-    let mut selection_ids = public_keys
-        .signers
-        .iter()
-        .map(|(&id, pk)| {
-            let pk_bytes = pk.to_bytes();
-            let mut buffer =
-                Vec::with_capacity(pk_bytes.len() + stacks_tip_consensus_hash.as_bytes().len());
-            buffer.extend_from_slice(&pk_bytes[..]);
-            buffer.extend_from_slice(stacks_tip_consensus_hash.as_bytes());
-            let digest = Sha256Sum::from_data(&buffer).as_bytes().to_vec();
-            (digest, id)
-        })
-        .collect::<Vec<_>>();
-
-    // Sort the selection IDs based on the hash
-    selection_ids.sort_by_key(|(hash, _)| hash.clone());
-
-    // Get the first ID from the sorted list and retrieve its public key,
-    // or default to the first signer if none are found
-    selection_ids
-        .first()
-        .and_then(|(_, id)| public_keys.signers.get(id).map(|pk| (*id, *pk)))
-        .unwrap_or((0, public_keys.signers.get(&0).cloned().unwrap()))
-}
-
-#[cfg(test)]
-mod tests {
-    use std::fmt::Write;
-    use std::net::TcpListener;
-    use std::thread::{sleep, spawn};
-
-    use rand::distributions::Standard;
-    use rand::{thread_rng, Rng};
-    use rand_core::RngCore;
-
-    use super::*;
-    use crate::client::tests::{
-        build_get_peer_info_response, generate_stacks_node_info, write_response, MockServerClient,
-    };
-
-    fn generate_random_consensus_hash() -> String {
-        let rng = rand::thread_rng();
-        let bytes: Vec<u8> = rng.sample_iter(Standard).take(20).collect();
-        let hex_string = bytes.iter().fold(String::new(), |mut acc, &b| {
-            write!(&mut acc, "{:02x}", b).expect("Error writing to string");
-            acc
-        });
-        hex_string
-    }
-
-    fn mock_stacks_client_response(mock_server: TcpListener, random_consensus: bool) {
-        let consensus_hash = match random_consensus {
-            true => generate_random_consensus_hash(),
-            false => "64c8c3049ff6b939c65828e3168210e6bb32d880".to_string(),
-        };
-
-        println!("{}", consensus_hash);
-        let stacks_tip_height = thread_rng().next_u64();
-        let response = build_get_peer_info_response(stacks_tip_height, consensus_hash);
-
-        spawn(move || {
-            write_response(mock_server, response.as_bytes());
-        });
-        sleep(Duration::from_millis(100));
-    }
-
-    #[test]
-    fn calculate_coordinator_should_produce_unique_results() {
-        let number_of_tests = 5;
-        let generated_public_keys = generate_stacks_node_info(10, 4000, None).0.public_keys;
-        let mut results = Vec::new();
-
-        for _ in 0..number_of_tests {
-            let mock = MockServerClient::new();
-            mock_stacks_client_response(mock.server, true);
-
-            let (coordinator_id, coordinator_public_key) =
-                calculate_coordinator(&generated_public_keys, &mock.client);
-
-            results.push((coordinator_id, coordinator_public_key));
-        }
-
-        // Check that not all coordinator IDs are the same
-        let all_ids_same = results.iter().all(|&(id, _)| id == results[0].0);
-        assert!(!all_ids_same, "Not all coordinator IDs should be the same");
-
-        // Check that not all coordinator public keys are the same
-        let all_keys_same = results
-            .iter()
-            .all(|&(_, key)| key.key.data == results[0].1.key.data);
-        assert!(
-            !all_keys_same,
-            "Not all coordinator public keys should be the same"
-        );
-    }
-    fn generate_test_results(random_consensus: bool, count: usize) -> Vec<(u32, ecdsa::PublicKey)> {
-        let mut results = Vec::new();
-        let generated_public_keys = generate_stacks_node_info(10, 4000, None).0.public_keys;
-        for _ in 0..count {
-            let mock = MockServerClient::new();
-            mock_stacks_client_response(mock.server, random_consensus);
-            let result = calculate_coordinator(&generated_public_keys, &mock.client);
-            results.push(result);
-        }
-        results
-    }
-
-    #[test]
-    fn calculate_coordinator_results_should_vary_or_match_based_on_hash() {
-        let results_with_random_hash = generate_test_results(true, 5);
-        let all_ids_same = results_with_random_hash
-            .iter()
-            .all(|&(id, _)| id == results_with_random_hash[0].0);
-        let all_keys_same = results_with_random_hash
-            .iter()
-            .all(|&(_, key)| key.key.data == results_with_random_hash[0].1.key.data);
-        assert!(!all_ids_same, "Not all coordinator IDs should be the same");
-        assert!(
-            !all_keys_same,
-            "Not all coordinator public keys should be the same"
-        );
-
-        let results_with_static_hash = generate_test_results(false, 5);
-        let all_ids_same = results_with_static_hash
-            .iter()
-            .all(|&(id, _)| id == results_with_static_hash[0].0);
-        let all_keys_same = results_with_static_hash
-            .iter()
-            .all(|&(_, key)| key.key.data == results_with_static_hash[0].1.key.data);
-        assert!(all_ids_same, "All coordinator IDs should be the same");
-        assert!(
-            all_keys_same,
-            "All coordinator public keys should be the same"
-        );
     }
 }
